requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mysql==0.0.3
mysqlclient==2.1.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.7
pytest==7.1.1
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.32
tomli==2.0.1
Werkzeug==2.0.3
